requirements.txt
opencv-python==4.10.0.84
Levenshtein==0.26.0
pillow==10.4.0
tqdm==4.66.5
filesplit==3.0.2
pysrt==1.1.2
wordsegment==1.3.1
scikit-image==0.25.1
lmdb==1.5.1
pyclipper==1.3.0.post5
pyside6-fluent-widgets==1.7.7
numpy==2.2.4
shapely==2.0.7
six==1.17.0
setuptools==75.8.0
paddleocr==2.10.0